services/diary-service/app/requirements.txt
flask==3.0.3
gunicorn==22.0.0
pymysql==1.1.1
werkzeug==3.0.3
python-dotenv==1.0.1
cryptography==42.0.8
